requirements.txt
asgiref==3.9.1
attrs==25.3.0
Django==5.2.4
django-dotenv==1.4.2
django-redis==6.0.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
packaging==25.0
psycopg2-binary==2.9.10
redis==6.2.0
typing_extensions==4.14.1
